Minor fixes (#1680)

+++ src/xma/include/lib/xmaxclbin.h
@@ -70,5 +70,5 @@ typedef struct XmaXclbinInfo
 
 char *xma_xclbin_file_open(const char *xclbin_name);
 int xma_xclbin_info_get(char *buffer, XmaXclbinInfo *info);
-int xma_xclbin_map2ddr(uint16_t bit_map, int* ddr_bank);
+int xma_xclbin_map2ddr(uint64_t bit_map, int* ddr_bank);
 #endif
